Move project tests into their own file
